+++ src/mapster.nim
@@ -1,5 +1,5 @@
-import ./mapster/mapster
-import ./mapster/mapsterVariant
+import ./mapster/map
+import ./mapster/mapVariant
 
-export mapster
-export mapsterVariant
+export map
+export mapVariant